requirements.txt
attrs==23.1.0
aws-cdk-lib==2.89.0
aws-cdk.asset-awscli-v1==2.2.200
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v5==2.0.166
aws-cdk.aws-lambda-python-alpha==2.89.0a0
cattrs==23.1.2
constructs==10.2.69
importlib-resources==6.0.0
jsii==1.85.0
psycopg2-binary==2.9.6
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
typeguard==2.13.3
typing_extensions==4.7.1
